Moving the InjuryLevel enum to the personnel.enums folder
