+++ MekHQ/src/mekhq/campaign/mod/am/InjuryTypes.java
@@ -32,7 +32,7 @@ import mekhq.campaign.log.MedicalLogEntry;
 import mekhq.campaign.log.MedicalLogger;
 import mekhq.campaign.personnel.enums.BodyLocation;
 import mekhq.campaign.personnel.Injury;
-import mekhq.campaign.personnel.InjuryLevel;
+import mekhq.campaign.personnel.enums.InjuryLevel;
 import mekhq.campaign.personnel.InjuryType;
 import mekhq.campaign.personnel.Modifier;
 import mekhq.campaign.personnel.Person;
