+++ MekHQ/src/mekhq/campaign/personnel/Injury.java
@@ -39,6 +39,7 @@ import javax.xml.bind.annotation.XmlRootElement;
 import javax.xml.bind.annotation.adapters.XmlJavaTypeAdapter;
 
 import mekhq.campaign.personnel.enums.BodyLocation;
+import mekhq.campaign.personnel.enums.InjuryLevel;
 import org.joda.time.DateTime;
 import org.w3c.dom.Node;
