+++ MekHQ/src/mekhq/gui/dialog/MedicalViewDialog.java
@@ -77,7 +77,7 @@ import mekhq.campaign.log.LogEntry;
 import mekhq.campaign.force.Force;
 import mekhq.campaign.personnel.enums.BodyLocation;
 import mekhq.campaign.personnel.Injury;
-import mekhq.campaign.personnel.InjuryLevel;
+import mekhq.campaign.personnel.enums.InjuryLevel;
 import mekhq.campaign.personnel.InjuryType;
 import mekhq.campaign.personnel.Person;
